Fixes serialization for interface changes
The interface changes introduced a FunctionOperation derived type
from Operation, which ends up initializing different operations
on different performers. Since this results in interface serializaing
differently on different performers, it had to be fixed.
