+++ base/src/interface/Plato_Operation.cpp
@@ -236,59 +236,6 @@ getOperationName() const
 }
 
 /******************************************************************************/
-void
-Operation::
-initializeBaseSingle(const Plato::OperationInputDataMng & aOperationDataMng,
-           const std::shared_ptr<Plato::Performer> aPerformer,
-           const std::vector<Plato::SharedData*>& aSharedData)
-/******************************************************************************/
-{
-    m_performer = nullptr;
-    m_parameters.clear();
-    m_inputData.clear();
-    m_outputData.clear();
-    m_argumentNames.clear();
-
-    m_performerName = aOperationDataMng.getPerformerName();
-    m_operationName = aOperationDataMng.getOperationName(m_performerName);
-
-    auto tAllParamsData = aOperationDataMng.get<Plato::InputData>("Parameters");
-    if( tAllParamsData.size<Plato::InputData>(m_performerName) )
-    {
-        auto tParamsData = tAllParamsData.get<Plato::InputData>(m_performerName);
-        for( auto tParamData : tParamsData.getByName<Plato::InputData>("Parameter") )
-        {
-            auto tArgName  = Plato::Get::String(tParamData,"ArgumentName");
-            auto tArgValue = Plato::Get::Double(tParamData,"ArgumentValue");
-            m_parameters.insert(
-              std::pair<std::string, Parameter*>(tArgName, new Parameter(tArgName, m_operationName, tArgValue)));
-        }
-    }
-
-    // Get the input shared data.
-    const int tNumInputs = aOperationDataMng.getNumInputs(m_performerName);
-    for(int tInputIndex = 0; tInputIndex < tNumInputs; tInputIndex++)
-    {
-        const std::string & tArgumentName = aOperationDataMng.getInputArgument(m_performerName, tInputIndex);
-        const std::string & tSharedDataName = aOperationDataMng.getInputSharedData(m_performerName, tInputIndex);
-        this->addArgument(tArgumentName, tSharedDataName, aSharedData, m_inputData);
-    }
-    
-    // Get the output shared data.
-    const int tNumOutputs = aOperationDataMng.getNumOutputs(m_performerName);
-    for(int tOutputIndex = 0; tOutputIndex < tNumOutputs; tOutputIndex++)
-    {
-        const std::string & tArgumentName = aOperationDataMng.getOutputArgument(m_performerName, tOutputIndex);
-        const std::string & tSharedDataName = aOperationDataMng.getOutputSharedData(m_performerName, tOutputIndex);
-        this->addArgument(tArgumentName, tSharedDataName, aSharedData, m_outputData);
-    }
-    if(aPerformer->myName() == m_performerName)
-    {
-        m_performer = aPerformer;
-    }
-}
-
-/******************************************************************************/
 void Operation::setPerformer(std::shared_ptr<Performer> aPerformer)
 /******************************************************************************/
 { 
@@ -296,6 +243,7 @@ void Operation::setPerformer(std::shared_ptr<Performer> aPerformer)
     if(m_performerName == aPerformer->myName())
     {
         m_performer = std::move(aPerformer);
+        setComputeFunctionOnNewPerformer();
     }
 }
